Update Main class

+++ src/main/java/house/House.java
@@ -6,7 +6,7 @@ public class House {
 
     private int id; ///id
     private int apartmentNumber; /// Номер квартири
-    private int area; /// Площа квартири
+    private double area; /// Площа квартири
     private int floor; /// поверх
     private int numberOfRooms; /// Кількість кімнат
     private String street; /// вулиця
@@ -20,7 +20,7 @@ public class House {
         street = "";
     }
 
-    public House(int id, int apartmentNumber, int area, int floor, int numberOfRooms, String street) {
+    public House(int id, int apartmentNumber, double area, int floor, int numberOfRooms, String street) {
         this.id = id;
         this.apartmentNumber = apartmentNumber;
         this.area = area;
@@ -32,14 +32,14 @@ public class House {
 
     public void SetId(int id){this.id = id;}
     public void SetApartmentNumber(int apartmentNumber){this.apartmentNumber = apartmentNumber;}
-    public void SetArea(int area){this.area = area;}
+    public void SetArea(double area){this.area = area;}
     public void SetFloor(int floor){this.floor = floor;}
     public void SetNumberOfRooms(int numberOfRooms){this.numberOfRooms = numberOfRooms;}
     public void SetStreet(String street){this.street = street;}
 
     public int GetId() {return id;}
     public int GetApartmentNumber() {return apartmentNumber;}
-    public int GetArea() {return area;}
+    public double GetArea() {return area;}
     public int GetFloor() {return floor;}
     public int GetNumberOfRooms() {return numberOfRooms;}
     public String GetStreet() {return street;}
